add new CustomTextView
