--- WildCat/WildCat/Controller/Pattern/PatternTableViewController.swift
@@ -1,116 +1,0 @@
-//
-//  PatternTableViewController.swift
-//  WildCat
-//
-//  Created by 陰山賢太 on 2018/10/23.
-//  Copyright © 2018 Azuma. All rights reserved.
-//
-
-import UIKit
-import Photos
-
-class PatternTableViewController: UITableViewController {
-
-    private var patterns:[Pattern] = []
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        self.setup()
-        self.navigationItem.rightBarButtonItem = UIBarButtonItem(barButtonSystemItem: .add, target: self, action: #selector(self.segueToAdd))
-        refreshControl = UIRefreshControl()
-        refreshControl?.addTarget(self, action: #selector(self.refreshData(sender:)), for: .valueChanged)
-    }
-
-    @objc func refreshData(sender: UIRefreshControl) {
-        refreshControl?.beginRefreshing()
-        let data = Array(Pattern.read())
-        self.patterns = data
-        self.tableView.reloadData()
-        refreshControl?.endRefreshing()
-    }
-
-    /// set Datas
-    private func setup() {
-        let data = Array(Pattern.read())
-        self.patterns = data
-        self.tableView.reloadData()
-    }
-
-    // MARK: - Table view data source
-
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        return 1
-    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return self.patterns.count
-    }
-
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "patternCell", for: indexPath) as! PatternTableViewCell
-        let targetPattern = self.patterns[indexPath.item]
-        var image = UIImage(named: "black")!
-        LocalPhoto.load(localIdentifer: targetPattern.imagePath) { (resultImage) in
-            if let resultImage = resultImage {
-                image = resultImage
-            }
-        }
-        cell.update(message: targetPattern.message, photo: image)
-        return cell
-    }
-
-    override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return CGFloat(105)
-    }
-
-    @objc func segueToAdd() {
-        let next = UIStoryboard(name: "PatternTableViewController", bundle: nil).instantiateViewController(withIdentifier: "add")
-        self.present(next, animated: true, completion: nil)
-    }
-
-    /*
-    // Override to support conditional editing of the table view.
-    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the specified item to be editable.
-        return true
-    }
-    */
-
-    /*
-    // Override to support editing the table view.
-    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete {
-            // Delete the row from the data source
-            tableView.deleteRows(at: [indexPath], with: .fade)
-        } else if editingStyle == .insert {
-            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-        }    
-    }
-    */
-
-    /*
-    // Override to support rearranging the table view.
-    override func tableView(_ tableView: UITableView, moveRowAt fromIndexPath: IndexPath, to: IndexPath) {
-
-    }
-    */
-
-    /*
-    // Override to support conditional rearranging of the table view.
-    override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the item to be re-orderable.
-        return true
-    }
-    */
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
