+++ WildCat/WildCat/Supporting View Controllers/CustomTextView.swift
@@ -1,0 +1,31 @@
+//
+//  CustomTextView.swift
+//  WildCat
+//
+//  Created by 山浦功 on 2018/10/27.
+//  Copyright © 2018 Azuma. All rights reserved.
+//
+
+import UIKit
+
+class CustomTextView: UITextView {
+
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        commonInit()
+    }
+
+    private func commonInit(){
+        let tools = UIToolbar()
+        tools.frame = CGRect(x: 0, y: 0, width: frame.width, height: 40)
+        let spacer = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: self, action: nil)
+        let closeButton = UIBarButtonItem(barButtonSystemItem: .done, target: self, action: #selector(self.closeButtonTapped))
+        tools.items = [spacer, closeButton]
+        self.inputAccessoryView = tools
+    }
+
+    @objc func closeButtonTapped(){
+        self.endEditing(true)
+        self.resignFirstResponder()
+    }
+}
